+++ src/main/java/org/folio/search/service/setter/instance/UniformTitleProcessor.java
@@ -10,7 +10,7 @@ import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.search.domain.dto.Instance;
-import org.folio.search.domain.dto.InstanceAlternativeTitles;
+import org.folio.search.domain.dto.InstanceAlternativeTitlesInner;
 import org.folio.search.integration.ReferenceDataService;
 import org.folio.search.service.setter.FieldProcessor;
 import org.springframework.stereotype.Component;
@@ -33,7 +33,7 @@ public class UniformTitleProcessor implements FieldProcessor<Instance, Set<Strin
 
     return toStreamSafe(instance.getAlternativeTitles())
       .filter(title -> uniformTitleIds.contains(title.getAlternativeTitleTypeId()))
-      .map(InstanceAlternativeTitles::getAlternativeTitle)
+      .map(InstanceAlternativeTitlesInner::getAlternativeTitle)
       .collect(toLinkedHashSet());
   }
 }